v0.1.4.1    (backup approaching v0.1.5)
basic log functionality | log is ArrayList that gets appended (to the
front) with each "entry" (event that adds a String to the log). First n
elements of log are then drawn to screen on each render frame | next
idea is to make each successive entry dimmer and dimmer. And have them
drop off the log after a certain amount of time.

participant is between entities and any other sub-class, ready for more
entities, which will as of now all fight between each other but not
amongst themselves | ideas to merge Participant and Entity, and methods
to set team; also a "team 66" clause in the attack() behavior that will
check if on team 66, and if they are, attack anyone except themselves.
Otherwise every entity will get its own team, and be set in the
constructor.

	modified:   out/linecounter.json
	modified:   out/linecounter.txt
	modified:   pocket/creature/Blue.java
	modified:   pocket/creature/Entity.java
	modified:   pocket/creature/Human.java
	modified:   pocket/creature/Participant.java
	modified:   pocket/creature/Red.java
	new file:   pocket/system/Counter.java
	new file:   pocket/system/Log.java
	modified:   pocket/system/Main.java
	new file:   pocket/system/Temp.java
	modified:   pocket/tile/Lava.java
	modified:   pocket/tile/Medkit.java
	modified:   pocket/world/Space.java
	new file:   save.txt