v0.1.6 "The Food Chain Update"
added food chain mechanics; creatures only eat when hungry, and will
resort to canibalism if they are starving. hunger increases over time
(30 sec. per unit, 5 min to full). Creatures will not attack up the
foodchain. a "food chain index" is used to specify where each creature
falls on the food chain. the index can vary, for example if a land
creature ends up in the water their index drops considerably, and vice
versa for water animals on land.

when things die now, instead of disappearing they turn into a corpse
item that can be picked up and eaten if hungry.

items exists, now, place them using [P]ickup Mode currently only swords,
shields, and corpses -- but more are soon to come. Swords and shields
don't currently give bonuses, but it will be as simple as just putting
in the details into the template I built. behaviors are automatically
executed for all pickups that are currently being held by someone.

log system is improved. now it's a Log Mode, access using [TAB], where
you can see up to a screen-full of the important events that have
occured; each entry includes details comprised of information that
varies depending on who was involved with the event

added a readme file with some basic instructions for the program

made it so cats wont voluntarily enter water, but they will exit it if
they happen to end up in it

added lava sharks!

	new file:   out/README
	modified:   out/linecounter.json
	modified:   out/linecounter.txt
	renamed:    save.txt -> out/save.txt
	new file:   pocket/creature/BigCat.java
	modified:   pocket/creature/Blue.java
	modified:   pocket/creature/Bug.java
	modified:   pocket/creature/Cat.java
	modified:   pocket/creature/Dog.java
	modified:   pocket/creature/Entity.java
	modified:   pocket/creature/Fish.java
	modified:   pocket/creature/Human.java
	modified:   pocket/creature/HumanAdult.java
	modified:   pocket/creature/LandAnimal.java
	new file:   pocket/creature/LavaShark.java
	modified:   pocket/creature/Lion.java
	modified:   pocket/creature/Man.java
	modified:   pocket/creature/Red.java
	modified:   pocket/creature/Rodent.java
	modified:   pocket/creature/Shark.java
	modified:   pocket/creature/Tiger.java
	modified:   pocket/creature/Woman.java
	new file:   pocket/pickup/Corpse.java
	new file:   pocket/pickup/Food.java
	new file:   pocket/pickup/Gun.java
	new file:   pocket/pickup/Medkit.java
	new file:   pocket/pickup/Pickup.java
	new file:   pocket/pickup/Shield.java
	new file:   pocket/pickup/Sword.java
	modified:   pocket/system/Dice.java
	modified:   pocket/system/Log.java
	modified:   pocket/system/Main.java
	new file:   pocket/system/Names.java
	modified:   pocket/system/Screen.java
	modified:   pocket/system/Temp.java
	new file:   pocket/tile/Fluid.java
	modified:   pocket/tile/Lava.java
	modified:   pocket/tile/Water.java
	modified:   pocket/world/Space.java
	modified:   pocket/world/World.java
	modified:   res/clear.png