v0.2
finally completed a save system. it took me days, and multiple
hours-long struggles trying to serialize a whole object. I tried putting
it in a string. I tried modifying Slick and using the modified.jar.
There was a rabbit-hole of things not being serializable. There were
problems with "circular references" and "deep nested graphs" that were
causing a stack over-flow. I tried switching to Kryo, which was supposed
 to be better able to handle the "deep-nested graphs". It wasn't, but it
recommended increasing the JVM stack size, which I did in vscode by
entering an additional command in the VMargs...

Anyway, in the end I went back to good 'ol "FileOutputStream,
ObjectOutputStream -- ...inputStream". After two or 3 days, I faced the
dragon head on. I was going to have to do this "old school" and save
only serializable data, like strings, ints and bools -- no packing whole
game states or objects into a file with "one magical line of code".

In the end I prevailed. The game saves exactly the way you'd want from a
user perspective, and the code is easy to work with. I wrote it so it's
as simple as calling save() and load() functions. How it works is that
the important data from each space in the world is copied into a
"mirror" object, or "ghost". The program placesa ghost in-front of each
space and copies the data into it. It only takes the data that is easily
serializable. The ghosts are then packed into an array called a "deck".
This "ghost deck", a sequence of mirror copies of every space in the
game is the big serializable object that is ultimately passed through
the IO Stream. Similarly on the loading side, we just unpack the ghost
deck, one ghost at a time, and write its data over every space in the
game.

	modified:   .vscode/launch.json
	modified:   .vscode/settings.json
	modified:   out/README
	modified:   out/linecounter.json
	modified:   out/linecounter.txt
	modified:   pocket/creature/BigCat.java
	modified:   pocket/creature/Blue.java
	modified:   pocket/creature/Bug.java
	modified:   pocket/creature/Cat.java
	renamed:    pocket/creature/Entity.java -> pocket/creature/Creature.java
	modified:   pocket/creature/Dog.java
	modified:   pocket/creature/Fish.java
	modified:   pocket/creature/Human.java
	modified:   pocket/creature/HumanAdult.java
	modified:   pocket/creature/LandAnimal.java
	modified:   pocket/creature/LavaShark.java
	modified:   pocket/creature/Lion.java
	modified:   pocket/creature/Man.java
	modified:   pocket/creature/Red.java
	modified:   pocket/creature/Rodent.java
	modified:   pocket/creature/Shark.java
	modified:   pocket/creature/Tiger.java
	modified:   pocket/creature/Woman.java
	new file:   pocket/item/Corpse.java
	renamed:    pocket/pickup/Food.java -> pocket/item/Food.java
	new file:   pocket/item/Gun.java
	renamed:    pocket/pickup/Pickup.java -> pocket/item/Item.java
	new file:   pocket/item/Medkit.java
	renamed:    pocket/pickup/Shield.java -> pocket/item/Shield.java
	renamed:    pocket/pickup/Sword.java -> pocket/item/Sword.java
	deleted:    pocket/pickup/Corpse.java
	deleted:    pocket/pickup/Gun.java
	deleted:    pocket/pickup/Medkit.java
	modified:   pocket/system/Cursor.java
	modified:   pocket/system/Dice.java
	new file:   pocket/system/Ghost.java
	modified:   pocket/system/Main.java
	new file:   pocket/system/MemoryCard.java
	modified:   pocket/system/Names.java
	modified:   pocket/system/Screen.java
	modified:   pocket/tile/Fluid.java
	modified:   pocket/tile/Lava.java
	modified:   pocket/tile/Medkit.java
	modified:   pocket/tile/Tile.java
	modified:   pocket/world/Space.java
	modified:   pocket/world/World.java